Implement tags and categories
The tag_and_category_generator.py script generates pages for each tag (in the
/tag directory) and category (in the /category directory). These files are just
short files with front matter directing them to use the /_layouts/tagpage or
/_layouts/categorypage layouts. If the front matter on all the pages needs to
be changed, update the Python script and run it again.

The post.html layout now includes the categories at the top, and the tags at
the bottom, with similar styling as the original site.

This site currently uses the remote broccolini swiss Jekyll theme, but when I
needed to overwrite it (e.g. change the post.html layout), I copied the remote
files into the relevant directories, so the site currently uses a mix of remote
and local files for the theme.